refactor: Like 구조 변경
