+++ src/main/java/com/spring/GreenJoy/domain/likes/LikeRepository.java
@@ -1,5 +1,6 @@
 package com.spring.GreenJoy.domain.likes;
 
+import com.spring.GreenJoy.domain.likes.entity.Like;
 import com.spring.GreenJoy.domain.post.entity.Post;
 import com.spring.GreenJoy.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
